fix get all items call
